+++ lib/galaxy_test/api/test_tools_upload.py
@@ -5,11 +5,11 @@ import urllib.parse
 import pytest
 from tusclient import client
 
-from galaxy.tool_util.unittest_utils import (
+from galaxy.tool_util.verify.test_data import TestDataResolver
+from galaxy.util.unittest_utils import (
     skip_if_github_down,
     skip_if_site_down,
 )
-from galaxy.tool_util.verify.test_data import TestDataResolver
 from galaxy_test.base.constants import (
     ONE_TO_SIX_ON_WINDOWS,
     ONE_TO_SIX_WITH_SPACES,
