+++ lib/galaxy_test/selenium/test_data_source_tools.py
@@ -1,4 +1,4 @@
-from galaxy.tool_util.unittest_utils import skip_if_site_down
+from galaxy.util.unittest_utils import skip_if_site_down
 from .framework import (
     managed_history,
     selenium_test,
